requirements.txt
# Core API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/LLM Dependencies
pyautogen==0.2.16
openai==1.12.0

# Web Search & Data Processing
tavily-python==0.3.3
requests==2.31.0

# Configuration & Environment
python-dotenv==1.0.0

# Frontend Dependencies
streamlit==1.28.1

# Data Visualization (for Docker execution)
matplotlib==3.8.2

# Utility Dependencies
docker==7.0.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1 